Simple posterior inference for beta-binomial when ncounts and ssize are only length 1.

Simple posterior inference for beta-binomial when ncounts and ssize are only length 1.

bb_simple_post(ncounts, ssize, ploidy, p1geno, p2geno, seq_error = 0,
  bias_val = 1, od_param = 0)

Arguments

ncounts
A vector of non-negative integers. The ith element is the number of counts of the ith sample.

ssize
A vector of positive integers. The ith element is the total number of counts of the ith sample.

ploidy
The ploidy of the species.

p1geno
The first parental genotype.

p2geno
The second parental genotype.

seq_error
A non-negative numeric. This is the known sequencing error rate. This is a rough high-ball error rate given by Li et. al. (2011).

bias_val
The bias parameter. A value of 1 means no bias.

od_param
The over-dispersion parameter. A value of 0 means no overdispersion. A value of 1 means total overdispersion.