Log-likelihood that flexdog maximizes when outliers are present.

Log-likelihood that flexdog maximizes when outliers are present.

flexdog_obj_out(probk_vec, out_prop, refvec, sizevec, ploidy, seq, bias,
  od, mean_bias, var_bias, mean_seq, var_seq)

Arguments

probk_vec
The kth element is the prior probability of genotype k (when starting to count from 0).

out_prop
The probability of being an outlier.

refvec
A vector of counts of reads of the reference allele.

sizevec
A vector of total counts.

ploidy
The ploidy of the species. Assumed to be the same for each individual.

seq
The starting value of the sequencing error rate.

bias
The starting value of the bias.

od
The starting value of the overdispersion parameter.

mean_bias
The prior mean of the log-bias.

var_bias
The prior variance of the log-bias.

mean_seq
The prior mean of the logit of the sequencing error rate.

var_seq
The prior variance of the logit of the sequencing error rate.

Value

A double. The flexdog objective when outliers = TRUE.

See also