Obtain the genotype distribution given the distribution of discrete uniforms.

Obtain the genotype distribution given the distribution of discrete uniforms.

get_probk_vec(pivec, model, mode)

Arguments

pivec
The mixing probability of the i'th discrete uniform distribution.

model
What form should the prior (genotype distribution) take? See Details for possible values.

mode
The mode if model = "ash". If not provided, flexdog will estimate the mode. This is the starting point of the allele frequency if model = "hw". This should be NULL for all other options of model.

Value

A vector of numerics. Element k is the probability of genotype k.

See also