E-step in flexdog where we now allow an outlier distribution.

E-step in flexdog where we now allow an outlier distribution.

get_wik_mat_out(probk_vec, out_prop, refvec, sizevec, ploidy, seq, bias, od)

Arguments

probk_vec
The vector of current prior probabilities of each genotype.

out_prop
The probability of being an outlier.

refvec
A vector of counts of reads of the reference allele.

sizevec
A vector of total counts.

ploidy
The ploidy of the species. Assumed to be the same for each individual.

seq
The starting value of the sequencing error rate.

bias
The starting value of the bias.

od
The starting value of the overdispersion parameter.

Value

Same as get_wik_mat but the last column is for the outlier class.

See also