Plot the beta distribution.

This function will plot the beta distribution under two different parameterizations. The first one, with alpha and beta, is the usual parameterization of the beta. The second one parameterizes the beta in terms of its mean and overdispersion. Thus, either both alpha and beta must be specified, or both mu and rho must be specified.

plot_beta_dist(alpha = NULL, beta = NULL, mu = NULL, rho = NULL, ...)

Arguments

alpha
The shape1 parameter.

beta
The shape2 parameter.

mu
The mean parameter.

rho
The overdispersion parameter.

...
Anything else you want to pass to plot.default.

Details

We have the relationships $$\mu = \alpha / (\alpha + \beta)$$ and $$\rho = 1 / (1 + \alpha + \beta).$$ Thus, the inverse relationships are $$\alpha = \mu(1 -\rho) / \rho$$ and $$\beta = (1 - \mu)(1 - \rho) / \rho.$$