Objective function when doing Brent's method in update_pp_f1 when one parent only has two mixing components.

Objective function when doing Brent's method in update_pp_f1 when one parent only has two mixing components.

pp_brent_obj(firstmixweight, probmat, pvec, weight_vec, alpha)

Arguments

firstmixweight
The mixing weight of the first component.

probmat
The rows index the components and the columns index the segregation amount. Should only have two rows.

pvec
The distribution of the other parent.

weight_vec
The weights for each element.

alpha
The mixing weight on the uniform component.

Value

The objective value, as calculated by taking a convolution using convolve of the mixing distribution and pvec, then putting that probability distribution through f1_obj.