Estimate SFS using algorithm of Li (2011)

We use the dynamic algorithm of Section 2.3.5 to obtain the site allele frequency likelihood for all SNPs, then use the EM algorithm of Section 2.3.8 to estimate the site frequency spectrum.

Usage

sfs_est(glarray, tol = 1e-05, itermax = 1000)

Arguments

glarray

An array of genotype log-likelihoods. Rows index SNPs, columns index individuals, slices index genotypes

tol

The stopping tolerance for the EM algorithm

itermax

The maximum number of EM iterations.

Value

The estimated SFS.

References

Li, H. (2011). A statistical framework for SNP calling, mutation discovery, association mapping and population genetical parameter estimation from sequencing data. Bioinformatics, 27(21), 2987-2993.

Author

David Gerard

Examples

# \donttest{
data("uitdewilligen", package = "updog")
mout <- multidog(
  refmat = uitdewilligen$refmat,
  sizemat = uitdewilligen$sizemat,
  ploidy = 4)
#>     |                                   *.#,%    
#>    |||                                 *******/  
#>  |||||||    (**..#**.                  */   **/  
#> |||||||||    */****************************/*%   
#>    |||    &****..,*.************************/    
#>    |||     (....,,,*,...****%********/(******    
#>    |||                ,,****%////,,,,./.****/    
#>    |||                  /**//         .*///....  
#>    |||                  .*/*/%#         .,/   ., 
#>    |||               , **/   #%         .*    .. 
#>    |||                               ,,,*        
#> 
#> Working on it...done!

## estimate SFS using genotype likelihoods
glarray <- format_multidog(x = mout, varname = paste0("logL_", 0:4))
sfs_gl <- sfs_est(glarray)

## compare to SFS using posterior modes
genomat <- format_multidog(x = mout, varname = "geno")
sfs_mode <- unclass(
  prop.table(
    table(
      factor(
        rowSums(genomat),
        levels = 0:(ncol(genomat) * 4)
      )
     )
   )
 )

## pretty similar, some differences:
df <- data.frame(mode = sfs_mode, gl = sfs_gl)
round(df[df$mode > 1e-6 | df$gl > 1e-6, ], digits = 3)
#>     mode    gl
#> 236  0.1 0.000
#> 237  0.0 0.100
#> 251  0.1 0.100
#> 257  0.1 0.144
#> 259  0.1 0.080
#> 261  0.2 0.189
#> 265  0.1 0.086
#> 266  0.0 0.000
#> 271  0.1 0.202
#> 272  0.1 0.000
#> 277  0.1 0.100
# }