Bayesian estimate of the SFS

This function will calculate the posterior distribution of the number of non-reference counts at each locus, then sum these probabilities over loci to get the posterior mean of the SFS.

Usage

sfs_pm(postarray)

Arguments

postarray

A 3-dimensional array of genotype posterior probabilities (not log-probabilities). The rows index the SNPs, the columns index the individuals, and the slices index the genotypes.

Value

The posterior mean SFS.

Author

David Gerard

Examples

# \donttest{
data("uitdewilligen", package = "updog")
mout <- multidog(
  refmat = uitdewilligen$refmat,
  sizemat = uitdewilligen$sizemat,
  ploidy = 4)
#>     |                                   *.#,%    
#>    |||                                 *******/  
#>  |||||||    (**..#**.                  */   **/  
#> |||||||||    */****************************/*%   
#>    |||    &****..,*.************************/    
#>    |||     (....,,,*,...****%********/(******    
#>    |||                ,,****%////,,,,./.****/    
#>    |||                  /**//         .*///....  
#>    |||                  .*/*/%#         .,/   ., 
#>    |||               , **/   #%         .*    .. 
#>    |||                               ,,,*        
#> 
#> Working on it...done!

postarray <- format_multidog(x = mout, varname = paste0("Pr_", 0:4))

sfs <- sfs_pm(postarray)

# }