Subset of individuals and SNPs from Uitdewilligen et al (2013).

A list containing a matrix of reference counts, a matrix of total counts, and the ploidy level (4) of the species. This is a subset of the data from Uitdewilligen et al (2013).

uitdewilligen

Format

A list containing three objects. Two matrices and a numeric scalar:

refmat

A matrix of read counts containing the reference allele. The rows index the individuals and the columns index the SNPs.

sizemat

A matrix of the total number of read counts. The rows index the individuals and the columns index the SNPs.

ploidy

The ploidy level of the species (just 4).

Source

https://doi.org/10.1371/journal.pone.0062355

Value

A list. See the Format Section.

References

Uitdewilligen, J. G., Wolters, A. M. A., Bjorn, B., Borm, T. J., Visser, R. G., & van Eck, H. J. (2013). A next-generation sequencing method for genotyping-by-sequencing of highly heterozygous autotetraploid potato. PLoS One, 8(5), e62355.

See also