Fixed-point iteration for updog EM.

Fixed-point iteration for updog EM.

up_fix(pival, p1geno, p2geno, alpha, beta, ocounts, osize, ldbinommat, qarray,
  r1vec, r2vec, update_pi = TRUE, update_geno = TRUE, update_beta = TRUE)

Arguments

pival
A fudge factor. The probability of not begin a mistake. Estimated if est_fudge = TRUE. Set to 0.99. Only used if est_fudge = TRUE.

p1geno
The genotype of parent 1

p2geno
The genotype of parent 2

alpha
The alpha parameter in the beta distribution of the outliers. Only used if est_fudge = TRUE.

beta
The beta parameter in the beta distribution of the outliers. Only used if est_fudge = TRUE.

ocounts
A vector of non-negative integers. The ith element is the number of reads of the reference allele in the ith child.

osize
A vector of positive integers. The ith element is the total number of reads for the ith child.

ldbinommat
The log binomial densities of the counts. The rows index the parental genotypes and the columns index the individuals.

qarray
An three-way array of proportions. The (i, j, k)th element is the probability of an offspring having k - 1 reference alleles given that parent 1 has i - 1 refrerence alleles and parent 2 has j - 1 reference alleles. Each dimension of the array is ploidy + 1.

r1vec
A vector of prior probabilities whose ith element is the prior probability that parent 1 has i -1 copies of allele A.

r2vec
A vector of prior probabilities whose ith element is the prior probability that parent 2 has i -1 copies of allele A.

update_pi
A logical. Should we update the fudge factor (TRUE) or not (FALSE)?

update_geno
A logical. Should we update the parental genotypes (TRUE) or not (FALSE)?

update_beta
A logical. Should we update the outlier model TRUE or not (FALSE)?