Same as update_pp_f1 but only allow s1.

Same as update_pp_f1 but only allow s1.

update_pp_s1(weight_vec, control)

Arguments

weight_vec
colSums(wik_mat) from flexdog. This is the sum of current posterior probabilities of each individual having genotype k.

control
A list of anything else needed to be passed. E.g. if model = "ash", then inner_weights needs to be passed through control (see get_inner_weights for how to get this matrix).

Value

A list with the following elements:

p1_pair_weights

A list with the mixing weights for the bivalent components of parent 1.

p2_pair_weights

A list with the mixing weights for the bivalent components of parent 2.

obj

The maximized objective.

p1geno

The estimated genotype for parent 1.

p2geno

The estimated genotype for parent 2.

pivec

The estimated genotype distribution for the offspring.

See also